Fixed form validation not listening to submit event.
