--- packages/inkline/src/mixins/methods/EmitSubmitMethodMixin.d.ts
@@ -1,5 +1,0 @@
-import { Component } from 'vue';
-
-declare const EmitSubmitMethodMixin: Component;
-
-export default EmitSubmitMethodMixin;
